+++ booking_web/src/components/pages/TodayBooking.jsx
@@ -1,50 +1,27 @@
-import React, { useState, useEffect } from "react";
+import React, { useEffect } from "react";
 import { useRecoilState, useSetRecoilState, useRecoilValue } from "recoil";
-import {
-  Box,
-  Stack,
-  Text,
-  Wrap,
-  WrapItem,
-} from "@chakra-ui/react";
+import { Box, Stack, Text, Wrap, WrapItem } from "@chakra-ui/react";
 
-import { today } from "../../utils/date";
-import { getAllBooking, getTables } from "../../utils/api";
 import { SelectedBooking } from "../organisms/SelectedBooking";
 import { NoAssignedBooking } from "../organisms/NoAssignedBooking";
 import { LeaveSeatButton } from "../atoms/button/LeaveSeatButton";
 import { SitSeatButton } from "../atoms/button/SitSeatButton";
 import { MainHeading } from "../atoms/heading/MainHeading";
 import {
-  todayBookingState,
   noAssignedBookingState,
   tableState,
   optionDispState,
 } from "../../atom/state";
+import { useFetchTodayBooking } from "../../hooks/useFetchTodayBooking";
 
 export const TodayBooking = () => {
-  const setTodayBooking = useSetRecoilState(todayBookingState);
   const [tables, setTables] = useRecoilState(tableState);
   const setOptionDisp = useSetRecoilState(optionDispState);
-  const [noAssigendBooking, setNoAssigendBooking] = useRecoilState(
-    noAssignedBookingState
-  );
+  const noAssigendBooking = useRecoilValue(noAssignedBookingState);
+  const { fetch } = useFetchTodayBooking();
 
   const fetchTodayBooking = async () => {
-    const allBooking = await getAllBooking();
-    const todayBooking = allBooking.filter(
-      (booking) => booking.date === today()
-    );
-    setTodayBooking(todayBooking);
-    const noAssigendBooking = todayBooking.filter(
-      (booking) => booking.tables.length === 0
-    );
-    setNoAssigendBooking(noAssigendBooking);
-    const tables = await getTables();
-    tables.sort((a, b) => {
-      return a.id - b.id;
-    });
-    setTables(tables);
+    fetch();
   };
 
   useEffect(() => {
@@ -63,7 +40,7 @@ export const TodayBooking = () => {
       <MainHeading>今日の予約</MainHeading>
       <Wrap p={{ base: 4, md: 6 }}>
         {tables.map((table) => (
-          <WrapItem key={table.id} >
+          <WrapItem key={table.id}>
             <Box
               w="200px"
               h="100%"
@@ -88,7 +65,7 @@ export const TodayBooking = () => {
       </Wrap>
       <Wrap p={{ base: 4, md: 6 }}>
         {noAssigendBooking.map((booking) => (
-          <WrapItem key={booking.id} >
+          <WrapItem key={booking.id}>
             <Box
               w="200px"
               h="100%"
